Create teletrap.conf
In order to encourage creative people to make more recordings, we need a setup that takes all the work out of using them. This new code generalizes what the helloyouthere.txt does and should make it easier to add/delete recordings, change transitions, add new series of clips etc.

With this include file, I have been able to implement not only what is done in helloyouthere.txt, but also the original Lenny script.

This file allows you to specify a group of recordings, optionally add random hellos for that group and specify the number of times to loop through the recordings before moving on to the next ones with optional transition recordings played between groups.